test: move to tests/bus.rs
